% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APutils.R
\name{add.AP}
\alias{add.AP}
\title{Fortify an DF with AP information}
\usage{
add.AP(df, Vvar)
}
\arguments{
\item{df}{Dataframe with the rawdata}

\item{Vvar}{name of the voltage variable in \code{df}, string}
}
\value{
source dataset with added AP index and AP centered timescale
}
\description{
This function takes an abf timeseries dataframe and detects AP within it.
Then it adds the index of the AP within the sweep and adds a timescale to align all AP found by their peaks.
}
