% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZAP.protocol.R
\name{calcZAP}
\alias{calcZAP}
\title{Calculation of complex impedance spectra from Current Clamp data}
\usage{
calcZAP(ZAPinput)
}
\arguments{
\item{ZAPinput}{A dataframe with timeseries data, V (voltage, mV) and I (current, pA) and sweep (works best using read multisweep before).
assumes 50khz sampling}
}
\value{
A dataframe with columns realZAP (MOhm) and imZAP, the real and imaginary complex impedance values at each frequency
}
\description{
This function takes a dataframe with columns V and I from current clamp
and calculates a impedance amplitude profile (ZAP).
For the formulae etc see \href{https://journals.physiology.org/doi/abs/10.1152/jn.1986.55.5.995}{this publication}
}
