% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.utils.R
\name{draw.first.sweep.png}
\alias{draw.first.sweep.png}
\title{Write a timeseries plot to file}
\usage{
draw.first.sweep.png(abf)
}
\arguments{
\item{abf}{An .abf file to create a plot of}
}
\value{
does not return a thing, just writes a png to file
}
\description{
This is the same function as "plot.first.sweep" but instead of returning the plot object it just writes a .png to file
in the directory of the .abf file.
This function is meant to be used in a l_ply call of a list of files.
}
