% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APutils.R
\name{findAP}
\alias{findAP}
\title{Find likely Peak points of Action potentials}
\usage{
findAP(dV)
}
\arguments{
\item{dv}{A vector of differentiated voltage values (dV/dt)}
}
\value{
A logical vector of length dV
}
\description{
This function checks a timeseries of differentiated voltages for whether the sign
of the derivative changes or it is 0.
Filtering for those points combined with filtering for overshoot (Voltage > 0) gave good results in identifying AP.
}
