% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.utils.R
\name{measure.timepoint}
\alias{measure.timepoint}
\title{Take a measurement of a timeseries at a given timewindow}
\usage{
measure.timepoint(var, tvar, tp, win)
}
\arguments{
\item{var}{the name of the variable}

\item{tvar}{the name of the time variable}

\item{tp}{the timepoint to measure}

\item{win}{the time window to take the measurement of, (centered on tp)}
}
\description{
This function takes a measurement at a given timepoint.
var being a vector in a dataframe with a time variable
tp being the timepoint, win being a window around it
this is meant for dplyr::summary calls so is supposed to return one datum
}
