% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.utils.R, R/load.multisweep.R
\name{plot.first.sweep}
\alias{plot.first.sweep}
\title{Create a timeseries graph from an abf}
\usage{
\method{plot}{first.sweep}(abf)

\method{plot}{first.sweep}(abf)
}
\arguments{
\item{abf}{An .abf file to create a plot of}
}
\value{
A ggplot object of the first sweep with voltage and current trace

A ggplot object of the first sweep with voltage and current trace
}
\description{
I dont have a way of browsing and displaying .abf on my linux machine.
Therefore this function just creates a plot of the first sweep so I can see what protocol the file is.
...Clampex misses the option of writing the protocol name to file...
Sampling rate defaults to 50000 Hz.

I dont have a way of browsing and displaying .abf on my linux machine.
Therefore this function just creates a plot of the first sweep so I can see what protocol the file is.
...Clampex misses the option of writing the protocol name to file...
Sampling rate defaults to 50000 Hz.
}
