% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passive.protocol.R
\name{process.passive}
\alias{process.passive}
\title{Process an abf file of the "passive" protocol}
\usage{
process.passive(abffile)
}
\arguments{
\item{abffile}{The File}
}
\value{
a .png and a .csv with the summarised analysis. Named with the filename trunk of 'abffile'.
}
\description{
This function takes an .abf file of the passive protocol and returns to the folder a summary graph and a .csv.
These contain per-sweep summaries.
SR of 50000 is assumed.
Function is tailored to this protocol only.
}
