% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZAP.protocol.R
\name{process.ZAP}
\alias{process.ZAP}
\title{Processing function for ZAP rawdata}
\usage{
process.ZAP(file, Vjunc)
}
\arguments{
\item{file}{ABF file with ZAP data (3 sweeps, 50000 hz SR)}

\item{Vjunc}{Junction potential to add}
}
\value{
writes to file ZAP csv file and plot png in the path of the input file
}
\description{
This function takes a path to a ZAP recording and writes the ZAP profile to file
Its intended to be used in l_ply calls
}
