% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.utils.R
\name{read.multisweep.custom}
\alias{read.multisweep.custom}
\title{Read a custom stimulus abf}
\usage{
read.multisweep.custom(file)
}
\arguments{
\item{abf}{periodic stimulation abf file}
}
\value{
tibble of the abf with stimulus and sweep variable
}
\description{
This is a wrapper around pyabf.
This function works on custom waveform stimulation (file-based) abf files.
The original protocol stimulus is reconstructed and added to the file.
This works with the ZAP protocol I designed which gets truncated by pclamp in a particular way, this reading function fixes that.
I have not tested with other protocols.
}
