% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.utils.R
\name{read.multisweep.pyth}
\alias{read.multisweep.pyth}
\title{Read .abf with the pyabf API}
\usage{
read.multisweep.pyth(file)
}
\arguments{
\item{abf}{periodic stimulation abf file}
}
\value{
tibble of the abf with stimulus and sweep variable
}
\description{
This is a wrapper around pyabf.
This function works on periodic stimulation abf files.
The original protocol stimulus is reconstructed and added to the file.
The function will fail on files with arbitrary waveform specification.
Here, the stimulus waveform file needs to be present in the project path or the directory of the rawdata file.
}
