% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.utils.R
\name{read.multisweep}
\alias{read.multisweep}
\title{Load a periodic stimulation abf file into a tidy data structure}
\usage{
read.multisweep(abf, SR)
}
\arguments{
\item{abf}{Path to the input .abf file (periodic abf, two timeseries)}

\item{SR}{The sampling rate of the recording in Hz}
}
\value{
A tibble object of the file with variables, time and sweep columns
}
\description{
This function takes an .abf file and loads it into memory as a tidy tibble.
Information about sweep etc is written to another column.
This will not always be a very memory saving way to do it.
This works for Vclamp and Cclamp recordings with two timeseries variables.
CAREFUL: This takes the recorded stimulus trace, not the applied stimulus ("add stimulus waveform" of clampfit)
Advantage of this function is that it should work with complex waveform protocols (AP waveform, ZAP) without the stimulus file being present.
}
