% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APutils.R
\name{returnAPdf}
\alias{returnAPdf}
\title{Fetch Action potentials}
\usage{
returnAPdf(df, Vvar, thresh1)
}
\arguments{
\item{df}{Time series df from an abf}

\item{Vvar}{Character vector, name of the voltage variable within df}

\item{thresh1}{threshold for detection of AP peaks, will default to 0 (aP is an AP if its overshoots)}
}
\value{
A data frame with variables described above, with list columns
}
\description{
This function takes a timeseries dataframe (from abf import) with the voltage variable \code{Vvar}.
It returns a dataframe containing the timepoints of the peaks of AP (\code{tpeak}), their height (\code{Vpeak}) and position within the sweep (\code{APindex}).
It contains the time scales \code{tAP} and \code{t} in list columns that can be matched with the source data timeseries to align AP for plotting.
}
