% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.utils.R
\docType{data}
\name{theme_linus}
\alias{theme_linus}
\title{Default setting for ggplot
Nice setting for ggplot}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 93.
}
\usage{
theme_linus
}
\description{
Default setting for ggplot
Nice setting for ggplot
}
\keyword{datasets}
