% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpulse.protocol.R
\name{process.VCtest}
\alias{process.VCtest}
\title{This function takes a membrane test pulse protocol and returns parameters such as
Input resistance capacitance and so on.}
\usage{
process.VCtest(abffile)
}
\arguments{
\item{abffile}{The File}
}
\value{
a .csv with the summarised analysis. Named with the filename trunk of 'abffile'.
}
\description{
This function takes a membrane test pulse protocol and returns parameters such as
Input resistance capacitance and so on.
}
