% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.utils.R
\name{read.multisweep.pyth}
\alias{read.multisweep.pyth}
\title{Read .abf with the pyabf API}
\usage{
read.multisweep.pyth(file)
}
\arguments{
\item{abf}{periodic stimulation abf file}
}
\value{
tibble of the abf with sweep variable
}
\description{
This is a wrapper around pyabf
}
